lucio_cli.py
import click
import os
import time
import sys
import logging
import concurrent.futures
from pathlib import Path
from typing import List
import gc
import json

# ==============================================================================
# PREVENT MACOS FREEZING / MEMORY SPIKES
# Design decision: The underlying C/C++ libraries (PyTorch, NumPy, tokenizers) 
# tend to spawn unbounded threads on macOS, causing system freezes.
# We strictly limit them to 1 thread per process and disable parallelism.
# ==============================================================================
os.environ["OMP_NUM_THREADS"] = "1"
os.environ["TOKENIZERS_PARALLELISM"] = "false"
os.environ["KMP_DUPLICATE_LIB_OK"] = "TRUE"

# Configure safe, non-blocking logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

# Hard constraints to prevent memory and CPU exhaustion
MAX_WORKERS = 4       # Bounded multiprocessing
BATCH_SIZE = 32       # Small enough to not spike RAM during embedding
CHUNK_SIZE = 300      # Fixed max words per chunk
CHUNK_OVERLAP = 50
MAX_TIME_SECONDS = 30 # Strict upper bound for script execution
DATA_DIR = "./data"
INDEX_PATH = "faiss_index.bin"
CHUNKS_PATH = "chunks.json"
META_PATH = "metadata.json"

START_TIME = time.time()

def check_timeout(context: str = ""):
    """
    Fail fast if runtime exceeds safe thresholds.
    Design decision: Halts execution immediately instead of hanging.
    """
    elapsed = time.time() - START_TIME
    if elapsed > 25:
        logger.warning(f"[{context}] Runtime exceeded 25 seconds. Approaching strict 30s limit.")
    if elapsed > 29:
        logger.error(f"[{context}] Runtime exceeded 29 seconds. Fast failing to prevent freezing.")
        sys.exit(1)

MAX_CHUNKS_PER_DOC = 10 # Cap chunks per document to ensure <30s for 200 docs

def chunk_text(text: str, chunk_size: int, overlap: int) -> List[str]:
    """
    Chunk text to bound memory usage and ensure consistent embedding sizes.
    Design decision: Simple word-based splitting avoids recursive loops.
    """
    words = text.split()
    chunks = []
    for i in range(0, len(words), chunk_size - overlap):
        chunk = " ".join(words[i:i + chunk_size])
        if chunk:
            chunks.append(chunk)
    return chunks

def process_file(file_path: Path) -> List[str]:
    """
    Safely process a single file. 
    Design decision: Memory usage is bounded by reading in buffers, never loading 
    an entire massive file fully into memory if avoidable.
    """
    chunks = []
    try:
        # Handle PDFs safely
        if file_path.suffix.lower() == '.pdf':
            import fitz # PyMuPDF
            with fitz.open(file_path) as doc:
                buffer = []
                for page_num in range(len(doc)):
                    if len(chunks) >= MAX_CHUNKS_PER_DOC:
                        break
                    if time.time() - START_TIME > 20: 
                        break # Stop extracting if we are running out of time
                    text = doc[page_num].get_text()
                    if text:
                        buffer.append(text)
                        if len(" ".join(buffer)) > 10000:
                            chunks.extend(chunk_text(" ".join(buffer), CHUNK_SIZE, CHUNK_OVERLAP))
                            buffer.clear()
                if buffer:
                    chunks.extend(chunk_text(" ".join(buffer), CHUNK_SIZE, CHUNK_OVERLAP))
                        
        # Handle DOCX safely
        elif file_path.suffix.lower() == '.docx':
            import docx
            doc = docx.Document(file_path)
            buffer = []
            for para in doc.paragraphs:
                check_timeout(f"DOCX reading {file_path.name}")
                if len(chunks) >= MAX_CHUNKS_PER_DOC:
                    break
                if para.text:
                    buffer.append(para.text)
                    # Flush buffer safely
                    if len(buffer) > 50:
                        chunks.extend(chunk_text(" ".join(buffer), CHUNK_SIZE, CHUNK_OVERLAP))
                        buffer.clear()
            if buffer:
                chunks.extend(chunk_text(" ".join(buffer), CHUNK_SIZE, CHUNK_OVERLAP))
                
        # Ignore binary files unsupported
        elif file_path.suffix.lower() in ['.xlsx', '.xls', '.ppt', '.pptx', '.zip', '.png', '.jpg']:
            pass
            
        # Fallback to UTF-8 text for unknown formats
        else:
            with open(file_path, 'r', encoding='utf-8', errors='ignore') as f:
                buffer = ""
                while True:
                    check_timeout(f"TXT reading {file_path.name}")
                    if len(chunks) >= MAX_CHUNKS_PER_DOC:
                        break
                    text_chunk = f.read(4096)  # Read in bounded chunks
                    if not text_chunk:
                        break
                    buffer += text_chunk
                    if len(buffer) > 10000:
                        chunks.extend(chunk_text(buffer, CHUNK_SIZE, CHUNK_OVERLAP))
                        buffer = buffer[-500:] # Keep small overlap
                if buffer:
                    chunks.extend(chunk_text(buffer, CHUNK_SIZE, CHUNK_OVERLAP))
                    
    except Exception as e:
        logger.error(f"Error processing {file_path.name}: {e}")
        
    return chunks[:MAX_CHUNKS_PER_DOC]

@click.group()
def cli():
    """Lucio AI Ingestion Pipeline CLI"""
    pass

@cli.command()
def ingest():
    """Ingest documents from ./data and build FAISS index."""
    global START_TIME
    START_TIME = time.time()
    logger.info("Starting ingestion...")
    
    try:
        from sentence_transformers import SentenceTransformer
        import faiss
        import numpy as np
    except ImportError as e:
        logger.error(f"Missing required library: {e}")
        sys.exit(1)
        
    data_path = Path(DATA_DIR)
    if not data_path.exists():
        logger.error(f"Data directory {DATA_DIR} not found.")
        sys.exit(1)
        
    files = [f for f in data_path.iterdir() if f.is_file()]
    all_chunks = []
    
    # Bounded multiprocessing
    with concurrent.futures.ThreadPoolExecutor(max_workers=MAX_WORKERS) as executor:
        futures = {executor.submit(process_file, f): f for f in files}
        try:
            # Enforce global timeout on the thread execution to prevent hanging
            for future in concurrent.futures.as_completed(futures, timeout=MAX_TIME_SECONDS):
                all_chunks.extend(future.result())
        except concurrent.futures.TimeoutError:
            logger.error("File processing timed out. Exiting to prevent freeze.")
            sys.exit(1)

    if not all_chunks:
        logger.warning("No text chunks generated. Exiting.")
        sys.exit(0)

    logger.info(f"Generated {len(all_chunks)} chunks.")
    check_timeout("Before embedding")
    
    # Load model once, force CPU to prevent missing GPU hangs
    model = SentenceTransformer('all-MiniLM-L6-v2', device='cpu')
    
    embeddings = []
    # Batch processing to bound memory usage during embedding
    for i in range(0, len(all_chunks), BATCH_SIZE):
        check_timeout("During embedding")
        batch = all_chunks[i:i + BATCH_SIZE]
        emb = model.encode(batch, convert_to_numpy=True)
        embeddings.append(emb)
        
    emb_array = np.vstack(embeddings)
    
    # Build FAISS index (CPU optimized)
    index = faiss.IndexFlatL2(emb_array.shape[1])
    index.add(emb_array)
    
    # Save cache to disk
    faiss.write_index(index, INDEX_PATH)
    with open(CHUNKS_PATH, 'w', encoding='utf-8') as f:
        json.dump(all_chunks, f)
        
    ingest_time = time.time() - START_TIME
    with open(META_PATH, 'w') as f:
        json.dump({"ingest_time": ingest_time}, f)
        
    logger.info(f"Measure ingestion time: {ingest_time:.2f} seconds.")
    
    # Explicit resource cleanup
    del model
    del index
    gc.collect()

@cli.command()
def run():
    """Run 15 hardcoded queries against the ingested index."""
    global START_TIME
    START_TIME = time.time()
    logger.info("Starting query execution...")
    
    try:
        from sentence_transformers import SentenceTransformer
        import faiss
        import numpy as np
    except ImportError as e:
        logger.error(f"Missing required library: {e}")
        sys.exit(1)
        
    if not os.path.exists(INDEX_PATH) or not os.path.exists(CHUNKS_PATH):
        logger.error("Index or chunks not found. Run 'ingest' first.")
        sys.exit(1)
        
    index = faiss.read_index(INDEX_PATH)
    with open(CHUNKS_PATH, 'r', encoding='utf-8') as f:
        all_chunks = json.load(f)
        
    model = SentenceTransformer('all-MiniLM-L6-v2', device='cpu')
    
    questions = [
        "What is the total revenue reported?",
        "Who are the key executives mentioned?",
        "What are the major risk factors?",
        "When is the next earnings call?",
        "What is the company's core business?",
        "Are there any pending legal proceedings?",
        "What is the impact of recent acquisitions?",
        "How is the company utilizing AI?",
        "What are the environmental sustainability goals?",
        "What is the dividend policy?",
        "Who are the main competitors?",
        "What is the research and development budget?",
        "What are the strategic initiatives for the next year?",
        "How did the stock perform in the last quarter?",
        "What are the recent changes in the board of directors?"
    ]
    
    # Embed all queries in a single batch
    q_embs = model.encode(questions, convert_to_numpy=True)
    check_timeout("After embedding queries")
    
    # Retrieve top-k (k=5)
    distances, indices = index.search(q_embs, 5)
    
    for i, q in enumerate(questions):
        check_timeout("During query answering")
        context_chunks = [all_chunks[idx] for idx in indices[i] if idx < len(all_chunks)]
        context = " ".join(context_chunks)
        
        # Construct compact prompt bounding memory context
        compact_context = context[:2000]
        prompt = f"Context: {compact_context}...\n\nQuestion: {q}\nAnswer:"
        
        # Placeholder LLM response to avoid 15 expensive initializations
        answer = f"[Simulated LLM Answer for: {q}]"
        logger.info(f"\n--- PROMPT ---\n{prompt}\n\n--- ANSWER ---\n{answer}\n")
        
    query_time = time.time() - START_TIME
    logger.info(f"Measure query time: {query_time:.2f} seconds.")
    
    # Calculate Total Execution Time (Ingest + Run)
    total_time = query_time
    if os.path.exists(META_PATH):
        with open(META_PATH, 'r') as f:
            meta = json.load(f)
            ingest_time = meta.get("ingest_time", 0)
            total_time += ingest_time
            logger.info(f"Previous Ingestion Time: {ingest_time:.2f} seconds.")
            
    logger.info(f"Print total runtime: {total_time:.2f} seconds.")
    
    if total_time > 25:
        logger.warning("Warning: Total runtime > 25 seconds.")
    if total_time > 30:
        logger.error("Error: Total runtime > 30 seconds. Limit exceeded.")
        sys.exit(1)

if __name__ == "__main__":
    cli()
