requirements.txt
click
sentence-transformers
faiss-cpu
numpy
pymupdf
python-docx
torch
